comments controller, tested
