+++ app/controllers/users_controller.rb
@@ -41,14 +41,6 @@ class UsersController < ApplicationController
     end
   end
 
-  def comment_post
-    @user = User.find(comment_params[:user_id])
-    @post = Post.find(comment_params[:post_id])
-    @user.comment(@post, comment_params[:body])
-
-    redirect_to post_path(@post)
-  end
-
   def friend_requests
     @user = User.find(params[:id])
     @requests = Friendship.requesters(@user)
@@ -120,9 +112,5 @@ class UsersController < ApplicationController
 
   def user_params
     params.require(:user).permit(:first_name, :last_name, :gender, :phone, :birthday)
-  end
-
-  def comment_params
-    params.require(:comment).permit(:body, :post_id, :user_id)
   end
 end
