+++ app/views/comments/_comment.html.erb
